+++ src/alwan/core/alwan_ipt_core.h
@@ -43,7 +43,7 @@ ALWAN_DIAG_POP
 
 ALWAN_DIAG_PUSH
 ALWAN_DIAG_DISABLE_FLOAT_CONV
-static const alwan_scalar IPT_V_EXPONENT = {
+ALWAN_CONSTEXPR alwan_scalar IPT_V_EXPONENT = {
 #include "../data/ipt_exponent.csv"
 };